// packages/realtime/src/proto/gen/auth/v1/auth_pb.ts
// @generated by protoc-gen-es v2.14.1 with parameter "target=ts"
// @generated from file auth/v1/auth.proto (package auth.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file auth/v1/auth.proto.
 */
export const file_auth_v1_auth: GenFile = /*@__PURE__*/
  fileDesc("ChJhdXRoL3YxL2F1dGgucHJvdG8SB2F1dGgudjEiEQoPU2Vzc2lvbnNDaGFuZ2VkIhEKD0pvaW5Db2RlQ2hhbmdlZCKDAQoKQXV0aFVwZGF0ZRI0ChBzZXNzaW9uc19jaGFuZ2VkGAEgASgLMhguYXV0aC52MS5TZXNzaW9uc0NoYW5nZWRIABI1ChFqb2luX2NvZGVfY2hhbmdlZBgCIAEoCzIYLmF1dGgudjEuSm9pbkNvZGVDaGFuZ2VkSABCCAoGdXBkYXRlQi5aLHRnYi1yZXNvbHZlci9zZXJ2ZXIvcHJvdG8vZ2VuL2F1dGgvdjE7YXV0aHYxYgZwcm90bzM");

/**
 * @generated from message auth.v1.SessionsChanged
 */
export type SessionsChanged = Message<"auth.v1.SessionsChanged"> & {
};

/**
 * Describes the message auth.v1.SessionsChanged.
 * Use `create(SessionsChangedSchema)` to create a new message.
 */
export const SessionsChangedSchema: GenMessage<SessionsChanged> = /*@__PURE__*/
  messageDesc(file_auth_v1_auth, 0);

/**
 * @generated from message auth.v1.JoinCodeChanged
 */
export type JoinCodeChanged = Message<"auth.v1.JoinCodeChanged"> & {
};

/**
 * Describes the message auth.v1.JoinCodeChanged.
 * Use `create(JoinCodeChangedSchema)` to create a new message.
 */
export const JoinCodeChangedSchema: GenMessage<JoinCodeChanged> = /*@__PURE__*/
  messageDesc(file_auth_v1_auth, 1);

/**
 * @generated from message auth.v1.AuthUpdate
 */
export type AuthUpdate = Message<"auth.v1.AuthUpdate"> & {
  /**
   * @generated from oneof auth.v1.AuthUpdate.update
   */
  update: {
    /**
     * @generated from field: auth.v1.SessionsChanged sessions_changed = 1;
     */
    value: SessionsChanged;
    case: "sessionsChanged";
  } | {
    /**
     * @generated from field: auth.v1.JoinCodeChanged join_code_changed = 2;
     */
    value: JoinCodeChanged;
    case: "joinCodeChanged";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message auth.v1.AuthUpdate.
 * Use `create(AuthUpdateSchema)` to create a new message.
 */
export const AuthUpdateSchema: GenMessage<AuthUpdate> = /*@__PURE__*/
  messageDesc(file_auth_v1_auth, 2);